requirements.txt
python-dotenv
requests
openai
fastapi
pydantic
uvicorn[standard]